Refactor common code in an abstract class.
