+++ src/API/Responses/Message.php
@@ -4,12 +4,12 @@ namespace madpilot78\bottg\API\Responses;
 
 use InvalidArgumentException;
 
-class Message implements ResponseObjectInterface
+class Message extends ResponseObject implements ResponseObjectInterface
 {
     /**
      * @var array
      */
-    private const FORMAT = [
+    protected $format = [
         'message_id'                => ['int', true],
         'from'                      => ['User', true], // May be an empty object, tranformed in null
         'date'                      => ['int', true],
@@ -95,52 +95,4 @@ class Message implements ResponseObjectInterface
      * @var string
      */
     public $text;
-
-    /**
-     * Populates object with data from relevant decoded reply part.
-     *
-     * @param object $src
-     *
-     * @throws InvalidArgumentException
-     *
-     * @return void
-     */
-    public function __construct(object $src)
-    {
-        foreach (self::FORMAT as $k => $v) {
-            if ($v[1] && !property_exists($src, $k)) {
-                throw new InvalidArgumentException('Required value missing: ' . $k);
-            } elseif (property_exists($src, $k)) {
-                $class = null;
-
-                // if $v[0] starts with an upper case it's a class name
-                if (ctype_upper(substr($v[0], 0, 1))) {
-                    $class = '\\madpilot78\\bottg\\API\\Responses\\' . $v[0];
-                    $valid = class_exists($class);
-                } else {
-                    $is_v = 'is_' . $v[0];
-                    $valid = $is_v($src->$k);
-                }
-
-                if (!$valid) {
-                    throw new InvalidArgumentException('Invalid value: ' . $k);
-                }
-
-                if (property_exists($src, $k)) {
-                    if (isset($class)) {
-                        // Empty properties are converted to null.
-                        $t = (array) $src->$k;
-
-                        if (empty($t)) {
-                            $this->$k = null;
-                        } else {
-                            $this->$k = new $class($src->$k);
-                        }
-                    } else {
-                        $this->$k = $src->$k;
-                    }
-                }
-            }
-        }
-    }
 }
